pyproject.toml
[project]
name = "ragiq"
version = "0.1.0"
description = "Production-grade RAG system with parsing, chunking, and retrieval"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "RAG IQ Team" }
]
license = { text = "MIT" }

dependencies = [
    # Core dependencies
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "tiktoken>=0.5.0",
    
    # Document parsing
    "pillow>=10.0.0",
    "pypdfium2>=4.30.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    
    # Chunking
    "llama-index-core>=0.10.0",
    "llama-index-readers-file>=0.1.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "markdown-it-py>=3.0.0",
    
    # Embedding and LLM
    "openai>=1.0.0",
    
    # Vector store
    "pinecone>=3.0.0",
    
    # UI
    "streamlit>=1.28.0",
    
    # Layout detection
    "doclayout-yolo>=0.0.4",
    "opencv-python-headless>=4.8.0",
    "huggingface-hub>=0.20.0",
    "onnxruntime>=1.16.0",
    "onnxtr[cpu]>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# ONNX runtime - platform specific
onnx = [
    "onnxruntime>=1.16.0; platform_machine == 'x86_64'",
    "onnxruntime-gpu>=1.16.0; platform_machine == 'x86_64'",
]

# Doctr OCR
doctr = [
    "onnxtr[gpu-headless]>=0.6.0; platform_machine == 'x86_64'",
    "onnxtr[cpu]>=0.6.0; platform_machine != 'x86_64'",
]

# Optional PyTorch for YOLO models
torch = [
    "torch>=2.0.0",
    "doclayout-yolo>=0.0.4",
    "opencv-python-headless>=4.8.0",
]

# Optional transformers for reading order
transformers = [
    "transformers>=4.30.0",
]

# Optional visualization
viz = [
    "plotly>=5.17.0",
    "matplotlib>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

